refactor: move customer/vendor memos into their domain constants.ts (out of constants/personal)
