+++ src/fixtures/customers/constants.ts
