+++ src/fixtures/vendors/constants.ts
